+++ Assets/Scripts/Composite/RawBonus.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 
 public class RawBonus : BaseAttribute {
     
-    public RawBonus(int value = 0, int multiplier = 0) : base(value, multiplier) {
+    public RawBonus(int value = 0, float multiplier = 0) : base(value, multiplier) {
         
     }
